requirements.txt
imageio==2.26.1
moviepy==1.0.3
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
streamlit==1.20.0
streamlit_option_menu==0.3.2
ultralytics==8.0.112
